+++ license-rtu-editor/src/app/license-rtu-editor/license-rtu-editor.component.ts
@@ -18,8 +18,9 @@
  * ===============LICENSE_END==================================================
  */
 
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild, Input } from '@angular/core';
 import { LicenseRtuMetaService } from '../license-rtu-meta.service';
+import { JsonSchemaFormComponent } from '@earlyster/angular6-json-schema-form';
 
 @Component({
   selector: 'app-license-rtu-editor',
@@ -41,16 +42,80 @@ export class LicenseRtuEditorComponent implements OnInit {
   downloadType = 'txt';
   jsonFormOptions: any;
   isValid: any;
+  rtuEditorForm: JsonSchemaFormComponent;
+  queryParams: any = {};
+
+  @Input() mode: string;
 
 
   constructor(private service: LicenseRtuMetaService) { }
 
+  /**
+   * JavaScript Get URL Parameter parsing
+   *
+   * This is utility function defined here but can be movied to
+   * Utils.ts if more usage required.
+   */
+  initQueryParams() {
+    const search = decodeURIComponent( window.location.href.slice( window.location.href.indexOf( '?' ) + 1 ) );
+    const definitions = search.split( '&' );
+
+    definitions.forEach((val) => {
+      const parts = val.split('=', 2);
+      this.queryParams[parts[0]] = parts[1];
+    });
+  }
+
+  // addEventListener and old browser support
+  bindEvent(element, eventName, eventHandler) {
+    if (element.addEventListener) {
+        element.addEventListener(eventName, eventHandler, false);
+    } else if (element.attachEvent) {
+        element.attachEvent('on' + eventName, eventHandler);
+    }
+  }
+
+  initIframeSetup() {
+    // Listen to messages from parent window
+    this.bindEvent(window, 'message', (event) => {
+      if (event.data.key === 'input') {
+        this.jsonData = event.data.value;
+      }
+    });
+  }
+
+  // Send a message to the parent
+  sendMessage(msgObj) {
+    window.parent.postMessage(msgObj, '*');
+  }
+
   ngOnInit() {
+
+    // attempt to read mode value from the query param, if not given as input
+    if (!this.mode) {
+      this.initQueryParams();
+      if (this.queryParams.mode) {
+        this.mode = this.queryParams.mode;
+      }
+    }
+    if (this.mode === 'iframe') {
+      this.initIframeSetup();
+    }
+
     // load assets/rtu-schema.json
     this.service.getSchema().subscribe((data) => {
       this.jsonSchema = data;
     });
 
+    this.jsonFormOptions = {
+      addSubmit: false, // Add a submit button if layout does not have one
+      debug: true,
+      setSchemaDefaults: true, // Always use schema defaults for empty fields
+      defautWidgetOptions: {
+        feedback: true, // Show inline feedback icons
+        listItems: 0 // Number of list items to initially add to arrays with no default value
+      }
+    };
 
     this.formLayout = [
       { type: 'flex', 'flex-flow': 'row wrap' },
@@ -86,10 +151,35 @@ export class LicenseRtuEditorComponent implements OnInit {
             {
               key: 'permission[].action',
               type: 'array',
+              listItems: 1,
               items: [
+                // adding the @type with condition: false as there is layout issue in schema editor code
+                // - layout:
+                //   For the FormArray type, the layout system creates a fieldset if the list item has
+                //   MORE than one property to display and under that fieldset it adds a FormGroup control
+                //   for each list item entry.
+                // - problem:
+                //   If the array item is configured to show single property then
+                //   the DOM structure is created without fieldset container for the child items.
+                //   Instead the FormGroup entry directly gets set as child item (with @type & action form controls)
+                //   - Clicking the Remove icon to remove the array item, removes the "action" form control
+                //     from the FormGroup parent and leaves the "@type" form control in parent FormGroup.
+                //     Now, as that FormControl remains/cached in FormArray, it results in
+                //     form validation error as the FormGroup has "@type" form control
+                //     but missing required "action" property.
+                // - solution:
+                //    Expected behavior: The FormGroup (with form controls for @type & action) array item entry
+                //    shall be removed from the parent FormArray.
+                //    To avoid above problem, adding the @type with "condition: false"
+                //    - (a) To keep the items entry > 1 (so that fieldset container gets added)
+                //    - (b) To keep the @type property hidden on screen
+                {
+                  key: 'permission[].action[].@type',
+                  condition: 'false'
+                },
                 {
                   key: 'permission[].action[].action'
-                },
+                }
               ]
             }, {
               key: 'permission[].constraint',
@@ -129,10 +219,15 @@ export class LicenseRtuEditorComponent implements OnInit {
             {
               key: 'prohibition[].action',
               type: 'array',
+              listItems: 1,
               items: [
                 {
-                  key: 'prohibition[].action[].action'
+                  key: 'prohibition[].action[].@type',
+                  condition: 'false'
                 },
+                {
+                  key: 'prohibition[].action[].action'
+                }
               ]
             }, {
               key: 'prohibition[].constraint',
@@ -173,6 +268,37 @@ export class LicenseRtuEditorComponent implements OnInit {
     });
   }
 
+  isValidFn(isValid) {
+    this.isValid = isValid;
+  }
+
+  @ViewChild(JsonSchemaFormComponent, { static: false })
+  set rtuEditorFormCmp(cmp: JsonSchemaFormComponent) {
+    if (!this.rtuEditorForm && cmp) {
+      this.rtuEditorForm = cmp;
+    }
+  }
+
+  saveRTU() {
+    const formData = this.rtuEditorForm.jsf.validData;
+    // - post license profile JSON data
+    this.sendMessage({
+      key: 'output',
+      value: formData
+    });
+  }
+
+  cancelRTU() {
+    this.sendMessage({
+      key: 'action',
+      value: 'cancel'
+    });
+  }
+
+  async downloadRTU() {
+    const formData = this.rtuEditorForm.jsf.validData;
+    this.download(formData);
+  }
 
   // create a yaml (text) or json file from the json model
   async download(formData) {